refactor: MockData 분리
