+++ client/__mocks__/apis/images.ts
@@ -1,12 +1,13 @@
 import { rest } from 'msw';
 import { mockApiURL } from '@mocks/apis/utils';
+import { IMG_SRC } from '@mocks/data';
 
 export const mockedUpload = rest.post(
   mockApiURL('/images/upload'),
   (req, res, ctx) =>
     res(
       ctx.json({
-        url: 'https://avatars.githubusercontent.com/u/20358042?s=48&v=4',
+        url: IMG_SRC,
       }),
     ),
 );
