+++ client/__mocks__/apis/posts.ts
@@ -1,6 +1,6 @@
 import { rest } from 'msw';
 import { mockApiURL } from '@mocks/apis/utils';
-import { POST } from '@mocks/MockData';
+import { POST } from '@mocks/data/posts';
 
 export const mockWritePost = rest.post(mockApiURL('/posts'), (req, res, ctx) =>
   res(ctx.json(1)),
