+++ client/__mocks__/data/index.ts
@@ -1,0 +1,2 @@
+export const IMG_SRC =
+  'https://avatars.githubusercontent.com/u/20358042?s=48&v=4';
