+++ client/__stories__/components/common/Header/Header.stories.tsx
@@ -1,6 +1,7 @@
 import React from 'react';
 import Header from '@/components/common/Header/Header';
 import { ComponentStory } from '@storybook/react';
+import { USER, USER_LONG_NICKNAME } from '@mocks/data/users';
 
 export default {
   component: Header,
@@ -30,19 +31,11 @@ MyBOX.args = {
 export const WithUser = Template.bind({});
 WithUser.args = {
   ...Main.args,
-  user: {
-    nickname: 'test',
-    image: 'https://avatars.githubusercontent.com/u/20358042?s=48&v=4',
-    role: 'ROLE_USER',
-  },
+  user: USER,
 };
 
 export const 닉네임_긴_경우 = Template.bind({});
 닉네임_긴_경우.args = {
   ...Main.args,
-  user: {
-    nickname: '일이삼사오육칠팔구십일이삼사오육',
-    image: 'https://avatars.githubusercontent.com/u/20358042?s=48&v=4',
-    role: 'ROLE_USER',
-  },
+  user: USER_LONG_NICKNAME,
 };
